requirements.txt
telethon
fastapi
uvicorn
openai  # Optional, for LLM
asyncpg 
databases
sqlalchemy
ollama
chromadb
langchain
langchain-community
langchain-text-splitters
beautifulsoup4
GitPython